shifts common functions to utils.py
