+++ benchmarks/benchmarks/bench_wigner.py
@@ -1,4 +1,6 @@
-"""Benchmarks for Wigner-d recursions"""
+"""
+Benchmarks for Wigner-d recursions
+"""
 
 # set threads used by numpy (before numpy is imported!)
 # import os
@@ -8,11 +10,10 @@
 # os.environ["VECLIB_MAXIMUM_THREADS"] = "4" # export VECLIB_MAXIMUM_THREADS=4
 # os.environ["NUMEXPR_NUM_THREADS"] = "6" # export NUMEXPR_NUM_THREADS=6
 
-from functools import partial
-from itertools import product
+
 import argparse
 import numpy as np
-import timeit
+
 import pyssht as ssht
 import s2fft
 import jax
@@ -32,30 +33,13 @@ config.update("jax_enable_x64", True)  # this only works on startup!
 
 # list of different parameters to benchmark
 # harmonic band-limit
-L_VALUES = [16, 32]
+L_VALUES = [16]
 # colatitude
 BETA = np.pi / 2
 # harmonic order
 MM = 0
 
-
-def parametrize(parameter_dict):
-    """
-    Returns a function that unpacks a dictionary's keys and values
-    """
-
-    def decorator(function):
-        function.param_names = list(parameter_dict.keys())
-        function.params = list(parameter_dict.values())
-        return function
-
-    return decorator
-
-
-def parameters_string(parameters, names):
-    return (
-        "(" + ", ".join(f"{name}: {val}" for name, val in zip(names, parameters)) + ")"
-    )
+from utils import parametrize, parameters_string, run_benchmarks
 
 
 # Risbo
@@ -135,32 +119,6 @@ def turok_jax_compute_slice_largest_plane(L):
     Turok: compute slice of wigner-d plane for L - 1 at MM (jax)
     """
     dl = s2fft.wigner.turok_jax.compute_slice(BETA, L - 1, L, MM).block_until_ready()
-
-
-def run_benchmarks(benchmarks, number_runs, number_repeats, print_results=True):
-    results = {}
-    for benchmark in benchmarks:
-        results[benchmark.__name__] = {}
-        if print_results:
-            print(benchmark.__name__)
-        for parameters in product(*benchmark.params):
-            parameters_key = tuple(zip(benchmark.param_names, parameters))
-            results[benchmark.__name__][parameters_key] = {}
-            benchmark_function = partial(benchmark, *parameters)
-            run_times = [
-                time / number_runs
-                for time in timeit.repeat(
-                    benchmark_function, number=number_runs, repeat=number_repeats
-                )
-            ]
-            results[benchmark.__name__][parameters_key]["time"] = run_times
-            if print_results:
-                print(
-                    f"{parameters_string(parameters, benchmark.param_names):>40}: "
-                    f"min(time): {min(run_times):>#7.2g}s, "
-                    f"max(time): {max(run_times):>#7.2g}s"
-                )
-    return results
 
 
 if __name__ == "__main__":
